Distinguish between pictures and (&) in Inspect.
We now merge right-associated chains of (&), but only if the source
locations overlap.  We do not merge `pictures` at all.  In practice,
this provides a really nice result.  It does the right thing for all
of the following examples:

    a & b & c   -- merged in a chain

    (a & b) & c -- left-associated expressions are separate

    a & b where a = a1 & a2  -- a is a separate subexpression

    a & b where b = b1 & b2  -- b is a separate subexpression!!!
                             -- because there's no source overlap

    a & pictures([b, c])  -- pictures is separate from the & chain

It doesn't do the right thing for:

    a & (b & c)  -- accidentally merged into one chain.

But that's not a big deal at all.

See #725